requirements.txt
Flask
numpy
Pillow
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1+cpu
torchvision==0.20.1+cpu
prometheus_client
